Refactor refactor uri path to extension refactor email to extension refactor baseservice and Request >> send model to api
