+++ app/src/main/java/com/example/msigl62/coworkandroiduset/UiExtension.kt
@@ -2,9 +2,15 @@ package com.example.msigl62.coworkandroiduset
 
 import android.animation.Animator
 import android.animation.AnimatorListenerAdapter
+import android.content.Context
+import android.net.Uri
+import android.provider.MediaStore
+import android.support.v4.content.CursorLoader
 import android.view.View
 import android.view.ViewPropertyAnimator
 import android.widget.ImageView
+import java.util.regex.Matcher
+import java.util.regex.Pattern
 
 infix fun ImageView.load(url: String?) = this.apply {
     com.bumptech.glide.Glide.with(context).load(url).into(this)
@@ -15,6 +21,13 @@ fun View.simpleFadeInAnimation() = let {
             .alpha(1.0f)
             .setListener(null)
     this.show()
+}
+
+fun String?.emailPattern():Matcher{
+    val validEmail = "[a-zA-Z0-9\\+\\.\\_\\%\\-\\+]{1,256}" + "\\@" +
+            "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,64}" + "(" + "\\." +
+            "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,25}" + ")+"
+    return Pattern.compile(validEmail).matcher(this)
 }
 
 fun View.simpleFadeOutAnimation(): ViewPropertyAnimator = let {
@@ -36,4 +49,13 @@ fun View.show() = let {
     this.visibility = View.VISIBLE
 }
 
-
+fun Uri.getPath(context: Context): String? {
+    val arrData = arrayOf(MediaStore.Images.Media.DATA)
+    val loader = CursorLoader(context, this, arrData, null, null, null)
+    val cursor = loader.loadInBackground()
+    val columnIndex = cursor?.getColumnIndexOrThrow(MediaStore.Images.Media.DATA)
+    cursor?.moveToFirst()
+    val result = columnIndex?.let { cursor.getString(it) }
+    cursor?.close()
+    return result
+}
