+++ app/src/main/java/com/example/msigl62/coworkandroiduset/callapi/Request.kt
@@ -4,14 +4,17 @@ import android.util.Log
 import com.example.msigl62.coworkandroiduset.InterActor
 import com.example.msigl62.coworkandroiduset.model.Register
 import com.example.msigl62.coworkandroiduset.network.BaseRetrofit
+import com.google.gson.Gson
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.observers.DisposableObserver
 import io.reactivex.schedulers.Schedulers
 import retrofit2.Response
 
 class Request : InterActor.ActData {
+    private val gson = Gson()
     override fun requestVerify(user: Register, callback: InterActor.OnFinishRequest) {
-        BaseRetrofit.createRx()?.sendRequestVerify(user.facebookId, user.name, user.email, user.password, user.image)
+        val jsonConverter: String = gson.toJson(user)
+        BaseRetrofit.createRx()?.sendRequestVerify(jsonConverter)
                 ?.subscribeOn(Schedulers.io())
                 ?.observeOn(AndroidSchedulers.mainThread())
                 ?.subscribe(object : DisposableObserver<Response<Register>>() {
@@ -22,7 +25,7 @@ class Request : InterActor.ActData {
                     }
 
                     override fun onError(e: Throwable) {
-                        Log.e("throw wtf ",e.message)
+                        Log.e("throw wtf ", e.message)
                     }
                 })
     }
