+++ app/src/main/java/com/example/msigl62/coworkandroiduset/network/BaseRetrofit.kt
@@ -1,5 +1,6 @@
 package com.example.msigl62.coworkandroiduset.network
 
+import com.google.gson.GsonBuilder
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
@@ -9,9 +10,10 @@ import retrofit2.converter.gson.GsonConverterFactory
 class BaseRetrofit {
     companion object Factory {
         fun createRx(): BaseService? {
+            val gson = GsonBuilder().setLenient().create()
             val retrofit = Retrofit.Builder()
                     .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-                    .addConverterFactory(GsonConverterFactory.create())
+                    .addConverterFactory(GsonConverterFactory.create(gson))
                     .baseUrl(BaseUrl.baseUrl)
                     .client(setOkHttpClient())
                     .build()
