+++ app/src/main/java/com/example/msigl62/coworkandroiduset/network/BaseService.kt
@@ -2,7 +2,6 @@ package com.example.msigl62.coworkandroiduset.network
 
 import com.example.msigl62.coworkandroiduset.model.Register
 import io.reactivex.Observable
-import okhttp3.MultipartBody
 import retrofit2.Response
 import retrofit2.http.Multipart
 import retrofit2.http.POST
@@ -11,9 +10,5 @@ import retrofit2.http.Part
 interface BaseService {
     @Multipart
     @POST("index.php")
-    fun sendRequestVerify(@Part("facebook_id") facebook_id: String?,
-                          @Part("name") name: String?,
-                          @Part("email") email: String?,
-                          @Part("password") password: String?,
-                          @Part image: MultipartBody.Part?): Observable<Response<Register>>
+    fun sendRequestVerify(@Part("user") userData: String): Observable<Response<Register>>
 }
